ai-service/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
python-multipart
pytesseract
pdfplumber
Pillow
openai-whisper
google-generativeai
